style(contentful): move Contentful to its own class
to avoid circular dependencies
